Expose ref lines functions
