+++ lib/sparkline_svg/reference_line.ex
@@ -1,6 +1,7 @@
 defmodule SparklineSvg.ReferenceLine do
   @moduledoc false
 
+  alias SparklineSvg.Core
   alias SparklineSvg.ReferenceLine
 
   @type ref_line_opts :: %{
@@ -41,6 +42,63 @@ defmodule SparklineSvg.ReferenceLine do
       keys == [] -> {:ok, ref_lines}
       Enum.all?(keys, &valid_type?/1) -> {:ok, ref_lines}
       true -> {:error, :invalid_ref_line_type}
+    end
+  end
+
+  @spec max(Core.points()) :: Core.y()
+  def max(datapoints) do
+    {_x, y} = Enum.max_by(datapoints, fn {_x, y} -> y end)
+    y
+  end
+
+  @spec min(Core.points()) :: Core.y()
+  def min(datapoints) do
+    {_x, y} = Enum.min_by(datapoints, fn {_x, y} -> y end)
+    y
+  end
+
+  @spec avg(Core.points()) :: Core.y()
+  def avg(datapoints) do
+    {sum, count} =
+      Enum.reduce(datapoints, {0, 0}, fn {_x, y}, {sum, count} -> {sum + y, count + 1} end)
+
+    sum / count
+  end
+
+  @spec median(Core.points()) :: Core.y()
+  def median(datapoints) do
+    sorted_datapoints = Enum.sort_by(datapoints, fn {_x, y} -> y end)
+    length = Enum.count(sorted_datapoints)
+    mid = div(length, 2)
+
+    if rem(length, 2) == 0 do
+      {_x, left} = Enum.at(sorted_datapoints, mid - 1)
+      {_x, right} = Enum.at(sorted_datapoints, mid)
+
+      (left + right) / 2
+    else
+      {_x, y} = Enum.at(sorted_datapoints, mid)
+      y
+    end
+  end
+
+  @spec percentile(integer()) :: (Core.points() -> Core.y())
+  def percentile(nth) do
+    fn datapoints ->
+      values_count = length(datapoints)
+
+      case nth / 100 * (values_count + 1) do
+        n when n < 1 ->
+          0
+
+        n ->
+          sorted_values = Enum.map(datapoints, &elem(&1, 1)) |> Enum.sort()
+
+          case Enum.drop(sorted_values, max(0, trunc(n) - 1)) do
+            [a, b | _] -> a + (n - trunc(n)) * (b - a)
+            [a] -> a
+          end
+      end
     end
   end
